pyproject.toml
[project]
name = "aarnxen-multiai-mcp"
version = "0.1.0"
description = "Multi-AI MCP server with parallel consensus, caching, and cost tracking"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Abhishek", email = "abhishek@aarnxen.dev" }]
dependencies = [
    "mcp[cli]>=1.0.0",
    "openai>=1.55.0",
    "google-genai>=1.19.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "duckduckgo-search>=7.0.0",
]

[project.optional-dependencies]
vectors = ["sentence-transformers>=3.0.0"]

[project.scripts]
aarnxen = "aarnxen.server:main"
ax = "aarnxen.server:main"
aarnxen-dashboard = "aarnxen.dashboard:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]
